requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
openai==1.23.2
packaging==24.0
playwright==1.43.0
pluggy==1.5.0
pydantic==2.7.0
pydantic_core==2.18.1
pyee==11.1.0
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
text-unidecode==1.3
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
youtube-transcript-api==0.6.2
